+++ src/modules/texts/Text/components/Dialogs/DialogDeleteOneText/DialogDeleteOneText.tsx
@@ -3,7 +3,7 @@ import React, { memo } from "react";
 import { useContextText } from "modules/texts/Text/Context";
 import { useDialogText } from "modules/texts/Text/shared/hooks/useDialogText";
 import DialogDeleteOneTextView from "modules/texts/Text/components/Dialogs/DialogDeleteOneText/DialogDeleteOneTextView";
-import { fetchDeleteOneTextService } from "modules/texts/Text/components/Dialogs/DialogDeleteOneText/service";
+import { fetchDeleteOneTextService } from "modules/texts/Text/components/Dialogs/DialogDeleteOneText/services";
 import { TEXT_QUERY_KEY } from "shared/services/constants/texts";
 
 const DialogDeleteOneText: React.FC = () => {
