+++ src/modules/texts/Text/components/Dialogs/DialogDeleteOneText/services/index.ts
